Add files via upload
Task 1. Hough-Transform for line detection. Implement a second-derivative edge detector on the image
north.jpg and top_view.png (first convert to grayscale) using the Laplacian of Gaussian (LoG)
method described on page 724 (G&W book). Plot edges uses zero crossing. Also use a threshold
of 4% similar to Fig 10.22 (pg. 728, G&W) to plot a second edge image.
The goal of this step is to outline the lines correctly identify the block structure of top view
of IIT Jammu. 

Task 2: Perform Hough-Transform for circle detection on the image X.

Task 3: Hybrid Images: A hybrid image is a picture that combines the low-spatial frequencies of one
picture with the high spatial frequencies of another picture producing an image with an
interpretation that changes with viewing distance. The basic idea is that high frequency tends to
dominate at a closer distance, but as you move away only the low frequency (smooth) part of the
signal can be seen. By blending the high frequency portion of one image with the low-frequency
portion of another, you get a hybrid image that leads to different interpretations at different
distances.